rename test package to be in the same namespace as the actual package
