+++ src/test/java/fact/io/hdureader/ReadAuxFiles.java
@@ -1,9 +1,5 @@
-package fact.io.hdu_reader;
+package fact.io.hdureader;
 
-import fact.io.hdureader.BinTable;
-import fact.io.hdureader.BinTableReader;
-import fact.io.hdureader.FITS;
-import fact.io.hdureader.OptionalTypesMap;
 import org.junit.Test;
 
 import java.io.IOException;
