+++ src/test/java/fact/io/hdureader/ZFitsReaderTests.java
@@ -1,6 +1,5 @@
-package fact.io.hdu_reader;
+package fact.io.hdureader;
 
-import fact.io.hdureader.*;
 import org.junit.Test;
 
 import java.io.Serializable;
